+++ nens_auth_client/conf.py
@@ -5,6 +5,7 @@ from appconf import AppConf
 class NensAuthClientAppConf(AppConf):
     SCOPE = ["openid", "email", "profile"]
     STANDALONE = False
+    URL_NAMESPACE = "nens_auth_client:"  # prefixed to viewnames in reverse()
     AUTO_CREATE_REMOTE_USER = True
     TIMEOUT = 10  # Timeout for token, JWKS and discovery requests (seconds)
     LEEWAY = 120  # Amount of seconds that a token's expiry can be off