+++ nens_auth_client/urls.py
@@ -18,13 +18,14 @@ from django.conf import settings
 from django.contrib import admin
 from nens_auth_client import views
 
+from .apps import NensAuthClientConfig
 
 try:
     from django.urls import re_path
 except ImportError:  # Django 1.11 compatibility
     from django.conf.urls import url as re_path
 
-
+app_name = NensAuthClientConfig.name
 
 urlpatterns = [
     re_path("^authorize/$", views.authorize, name="authorize"),